requirements.txt
transformers>=4.40
torch>=2.0
datasets>=2.18
python-pptx
sentencepiece
openpyxl
safetensors